backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pandas
numpy
joblib
scikit-learn
pillow
tensorflow
librosa
soundfile
python-dotenv